video_processing_task/requirements.txt
amqp==5.3.1
billiard==4.2.1
boto3==1.35.92
botocore==1.35.92
botocore-stubs==1.35.96
celery==5.4.0
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
ffmpeg-python==0.2.0
future==1.0.0
gevent==24.11.1
greenlet==3.1.1
jmespath==1.0.1
kombu==5.4.2
loguru==0.7.3
prompt_toolkit==3.0.48
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
s3transfer==0.10.4
setuptools==75.7.0
six==1.17.0
types-awscrt==0.23.6
types-boto3==1.35.96
types-boto3-cloudformation==1.35.93
types-boto3-dynamodb==1.35.94
types-boto3-ec2==1.35.93
types-boto3-lambda==1.35.93
types-boto3-rds==1.35.95
types-boto3-s3==1.35.93
types-boto3-sqs==1.35.93
types-s3transfer==0.10.4
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
win32_setctime==1.2.0
zope.event==5.0
zope.interface==7.2